tests: rewrite runner in C
