+++ tests/runner/argparser.h
@@ -1,415 +1,219 @@
 #pragma once
 
-#include <vector>
-#include <unordered_map>
-#include <string>
-#include <functional>
-#include <algorithm>
-#include <stdexcept>
-#include <cstring>
-#include <memory>
-#include <iostream>
+#include "helpers.h"
+#include <stdbool.h>
+#include <string.h>
 
-class ArgParser
+typedef enum {
+    ARG_FLAG,
+    ARG_PARAM,
+    ARG_LIST,
+    ARG_HELP,
+    ARG_POSITIONAL,
+} arg_type_t;
+
+typedef struct {
+    const char *as_full;
+    char as_short;
+    arg_type_t type;
+    const char *description;
+    bool is_optional;
+    bool parsed;
+    vector_t values;
+} arg_spec_t;
+
+#define ARG_POS(name, desc)     \
+    {                           \
+        .as_full = name,        \
+        .type = ARG_POSITIONAL, \
+        .description = desc,    \
+    }
+#define ARG_LIST(name, short, desc) \
+    {                               \
+        .as_full = name,            \
+        .as_short = short,          \
+        .type = ARG_LIST,           \
+        .description = desc,        \
+        .is_optional = true,        \
+    }
+#define ARG_FLAG(name, short, desc) \
+    {                               \
+        .as_full = name,            \
+        .as_short = short,          \
+        .type = ARG_FLAG,           \
+        .description = desc,        \
+        .is_optional = true,        \
+    }
+#define ARG_PARAM(name, short, desc) \
+    {                                \
+        .as_full = name,             \
+        .as_short = short,           \
+        .type = ARG_PARAM,           \
+        .description = desc,         \
+        .is_optional = true,         \
+    }
+#define ARG_HELP(name, short, desc) \
+    {                               \
+        .as_full = name,            \
+        .as_short = short,          \
+        .type = ARG_HELP,           \
+        .description = desc,        \
+        .is_optional = true,        \
+    }
+
+typedef struct {
+    arg_spec_t *const *positional_args;
+    size_t num_positional_args;
+    arg_spec_t *const *option_args;
+    size_t num_option_args;
+} arg_parser_t;
+
+static inline void print_help(const arg_parser_t *parser)
 {
-public:
-    using help_callback = std::function<void()>;
+    size_t i;
 
-    ArgParser& add_param(
-        std::string_view full_arg, char short_arg, std::string_view description,
-        bool optional = true
-    )
-    {
-        return add_custom(full_arg, short_arg, ArgType::param, description,
-                          optional);
+    printf("uACPI test runner:\n");
+
+    for (i = 0; i < parser->num_positional_args; i++)
+        printf(
+            "           [%s] %s\n", parser->positional_args[i]->as_full,
+            parser->positional_args[i]->description
+        );
+
+    for (i = 0; i < parser->num_option_args; i++)
+        printf(
+            "%s [--%s/-%c] %s\n",
+            parser->option_args[i]->is_optional ? "(optional)" : "          ",
+            parser->option_args[i]->as_full,
+            parser->option_args[i]->as_short,
+            parser->option_args[i]->description
+        );
+}
+
+static inline bool is_arg(const char *arg)
+{
+    size_t length = strlen(arg);
+
+    switch (length) {
+    case 0:
+    case 1:
+        return false;
+    case 2:
+        return arg[0] == '-';
+    default:
+        return arg[0] == '-' && arg[1] == '-';
+    }
+}
+
+static inline void parse_args(
+    const arg_parser_t *parser, int argc, char *argv[]
+)
+{
+    size_t num_args = argc;
+    arg_spec_t *active_spec = NULL;
+    size_t arg_index;
+
+    if (num_args < 2) {
+        print_help(parser);
+        exit(1);
     }
 
-    ArgParser& add_flag(
-        std::string_view full_arg, char short_arg, std::string_view description,
-        bool optional = true
-    )
-    {
-        return add_custom(full_arg, short_arg, ArgType::flag, description,
-                          optional);
-    }
-
-    ArgParser& add_list(
-        std::string_view full_arg, char short_arg, std::string_view description,
-        bool optional = true
-    )
-    {
-        return add_custom(full_arg, short_arg, ArgType::list, description,
-                          optional);
-    }
-
-    ArgParser& add_help(
-        std::string_view full_arg, char short_arg,
-        std::string_view description, help_callback on_help_requested
-    )
-    {
-        m_help_callback = on_help_requested;
-        return add_custom(full_arg, short_arg, ArgType::help, description, true);
-    }
-
-    ArgParser& add_positional(
-        std::string_view name, std::string_view description)
-    {
-        if (m_num_positional_args != m_args.size()) {
-            throw std::runtime_error(
-                "positional argument follows keyword argument"
+    if (parser->num_positional_args) {
+        if ((num_args - 1) < parser->num_positional_args)
+            error(
+                "expected at least %zu positional arguments",
+                parser->num_positional_args
             );
-        }
 
-        m_num_positional_args++;
-        return add_custom(name, '\0', ArgType::positional, description, false);
+        for (arg_index = 0; arg_index < parser->num_positional_args;
+             ++arg_index)
+            vector_add(
+                &parser->positional_args[arg_index]->values,
+                argv[1 + arg_index], 0
+            );
     }
 
-    void parse(int argc, char** argv)
-    {
-        size_t num_args = argc;
+    for (arg_index = 1 + parser->num_positional_args; arg_index < num_args;
+         ++arg_index) {
+        char *current_arg = argv[arg_index];
+        bool is_new_arg = is_arg(current_arg);
+        arg_spec_t *new_spec = NULL;
+        size_t length;
 
-        if (num_args < 2) {
-            if (m_help_callback)
-                m_help_callback();
+        if (active_spec) {
+            if (!is_new_arg) {
+                if (active_spec->type == ARG_FLAG)
+                    error("unexpected argument %s", current_arg);
 
-            std::exit(1);
-        }
+                if (active_spec->type == ARG_PARAM &&
+                    active_spec->values.count == 1)
+                    error("too many arguments for %s", active_spec->as_full);
 
-        if (m_num_positional_args) {
-            if ((num_args - 1) < m_num_positional_args) {
-                throw std::runtime_error(
-                    "expected at least " + std::to_string(m_num_positional_args)
-                    + "positional arguments"
-                );
-            }
-
-            for (auto i = 0; m_parsed_args.size() < m_num_positional_args; ++i)
-                m_parsed_args[m_args[i].as_full].emplace_back(argv[1 + i]);
-        }
-
-        const ArgSpec* active_spec = nullptr;
-
-        for (auto arg_index = 1 + m_num_positional_args; arg_index < num_args;
-             ++arg_index) {
-            auto* current_arg = argv[arg_index];
-            auto is_new_arg = is_arg(current_arg);
-
-            if (active_spec) {
-                if (!is_new_arg) {
-                    if (active_spec->is_flag()) {
-                        throw std::runtime_error(
-                            std::string("unexpected argument ") + current_arg
-                        );
-                    }
-
-                    if (active_spec->is_param() &&
-                        m_parsed_args[active_spec->as_full].size() == 1) {
-                        throw std::runtime_error(
-                            "too many arguments for " + active_spec->as_full
-                        );
-                    }
-
-                    m_parsed_args[active_spec->as_full].emplace_back(
-                        current_arg
-                    );
-                    continue;
-                }
-
-                if ((active_spec->is_param() || active_spec->is_list()) &&
-                    m_parsed_args[active_spec->as_full].empty()) {
-                    throw std::runtime_error(
-                        "expected an argument for " + active_spec->as_full
-                    );
-                }
-            }
-
-            auto as_full_arg = extract_full_arg(current_arg);
-            if (as_full_arg.empty()) {
-                throw std::runtime_error(
-                    std::string("unexpected argument ") + current_arg
-                );
-            }
-
-            active_spec = &arg_spec_of(as_full_arg);
-            if (active_spec->is_help()) {
-                if (m_help_callback)
-                    m_help_callback();
-
-                std::exit(1);
-            }
-
-            m_parsed_args[as_full_arg];
-            continue;
-        }
-
-        ensure_mandatory_args_are_satisfied();
-    }
-
-    const std::vector<std::string>& get_list(std::string_view arg) const
-    {
-        ensure_arg_is_parsed(arg);
-
-        return m_parsed_args.at(std::string(arg));
-    }
-
-    const std::vector<std::string>& get_list_or(
-        std::string_view arg, const std::vector<std::string>& default_value
-    ) const
-    {
-        if (is_arg_parsed(arg))
-            return get_list(arg);
-
-        return default_value;
-    }
-
-    const std::string& get(std::string_view arg) const
-    {
-        return get_list(arg)[0];
-    }
-
-    const std::string& get_or(
-        std::string_view arg, const std::string& default_value
-    ) const
-    {
-        if (is_arg_parsed(arg))
-            return get(arg);
-
-        return default_value;
-    }
-
-    uint64_t get_uint(std::string_view arg) const
-    {
-        return std::stoull(get(arg));
-    }
-
-    uint64_t get_uint_or(std::string_view arg, uint64_t default_value) const
-    {
-        if (is_arg_parsed(arg))
-            return get_uint(arg);
-
-        return default_value;
-    }
-
-    int64_t get_int(std::string_view arg) const
-    {
-        return std::stoll(get(arg));
-    }
-
-    int64_t get_int_or(std::string_view arg, int64_t default_value) const
-    {
-        if (is_arg_parsed(arg))
-            return get_int(arg);
-
-        return default_value;
-    }
-
-    bool is_set(std::string_view arg) const
-    {
-        return is_arg_parsed(arg);
-    }
-
-    const std::vector<std::string>& get_list(char arg) const
-    {
-        const auto& arg_spec = arg_spec_of(arg);
-
-        return get_list(arg_spec.as_full);
-    }
-
-    std::string_view get(char arg) const
-    {
-        const auto& arg_spec = arg_spec_of(arg);
-
-        return get(arg_spec.as_full);
-    }
-
-    uint64_t get_uint(char arg) const
-    {
-        const auto& arg_spec = arg_spec_of(arg);
-
-        return get_uint(arg_spec.as_full);
-    }
-
-    uint64_t get_int(char arg) const
-    {
-        const auto& arg_spec = arg_spec_of(arg);
-
-        return get_int(arg_spec.as_full);
-    }
-
-    bool is_set(char arg) const
-    {
-        const auto& arg_spec = arg_spec_of(arg);
-
-        return is_set(arg_spec.as_full);
-    }
-
-    std::ostream& print(std::ostream& stream = std::cout) const
-    {
-        for (size_t i = 0; i < m_num_positional_args; ++i) {
-            auto& arg = m_args[i];
-            stream << "           [" << arg.as_full << "] " << arg.description
-                   << std::endl;
-        }
-
-        for (size_t i = m_num_positional_args; i < m_args.size(); ++i) {
-            auto& arg = m_args[i];
-
-            stream << (arg.is_optional ? "(optional) " : "           ")
-                   << "[--"
-                   << arg.as_full
-                   << "/-"
-                   << arg.as_short
-                   << "] "
-                   << arg.description
-                   << std::endl;
-        }
-
-        return stream;
-    }
-
-    friend std::ostream& operator<<(
-        std::ostream& stream, const ArgParser& argp
-    ) {
-        return argp.print(stream);
-    }
-
-private:
-    enum class ArgType
-    {
-        flag,
-        param,
-        list,
-        help,
-        positional,
-    };
-
-    struct ArgSpec
-    {
-        std::string as_full;
-        char        as_short;
-        ArgType     type;
-        std::string description;
-        bool        is_optional;
-
-        [[nodiscard]] bool is_list()  const { return type == ArgType::list; }
-        [[nodiscard]] bool is_param() const { return type == ArgType::param; }
-        [[nodiscard]] bool is_flag()  const { return type == ArgType::flag; }
-        [[nodiscard]] bool is_help()  const { return type == ArgType::help; }
-    };
-
-    void ensure_mandatory_args_are_satisfied() const
-    {
-        for (const auto& arg : m_args) {
-            if (arg.is_optional)
+                vector_add(&active_spec->values, current_arg, 0);
                 continue;
-
-            if (!m_parsed_args.count(arg.as_full))
-                throw std::runtime_error(
-                    "expected a non-optional argument --" + arg.as_full
-                );
-        }
-    }
-
-    ArgParser& add_custom(
-        std::string_view full_arg, char short_arg, ArgType type,
-        std::string_view description, bool optional
-    )
-    {
-        m_args.emplace_back(ArgSpec {
-            std::string(full_arg),
-            short_arg,
-            type,
-            std::string(description),
-            optional,
-        });
-
-        return *this;
-    }
-
-    bool is_arg_parsed(std::string_view arg) const
-    {
-        return m_parsed_args.count(std::string(arg));
-    }
-
-    void ensure_arg_is_parsed(std::string_view arg) const
-    {
-        if (!is_arg_parsed(arg)) {
-            throw std::runtime_error(
-                "couldn't find argument " + std::string(arg)
-            );
-        }
-    }
-
-    const ArgSpec& arg_spec_of(std::string_view arg) const
-    {
-        auto arg_itr = std::find_if(
-            m_args.begin() + m_num_positional_args,
-            m_args.end(),
-            [&](const ArgSpec& my_arg) {
-                return my_arg.as_full == arg;
             }
-        );
 
-        if (arg_itr == m_args.end())
-            throw std::runtime_error("unknown argument " + std::string(arg));
+            if ((active_spec->type == ARG_PARAM ||
+                active_spec->type == ARG_LIST) &&
+                active_spec->values.count == 0)
+                error("expected an argument for %s", active_spec->as_full);
+        }
 
-        return *arg_itr;
-    }
+        length = strlen(current_arg);
 
-    const ArgSpec& arg_spec_of(char arg) const
-    {
-        auto arg_itr = std::find_if(
-            m_args.begin() + m_num_positional_args, m_args.end(),
-            [&](const ArgSpec& my_arg) {
-                return my_arg.as_short == arg;
+        if (length >= 2) {
+            size_t i;
+
+            for (i = 0; i < parser->num_option_args; i++) {
+                arg_spec_t *spec = parser->option_args[i];
+
+                if (length == 2 && spec->as_short == current_arg[1]) {
+                    new_spec = spec;
+                    break;
+                } else if (strcmp(spec->as_full, &current_arg[2]) == 0) {
+                    new_spec = spec;
+                    break;
+                }
             }
-        );
-
-        if (arg_itr == m_args.end())
-            throw std::runtime_error(std::string("unknown argument ") + arg);
-
-        return *arg_itr;
-    }
-
-    bool is_arg(const char* arg)
-    {
-        size_t length = strlen(arg);
-
-        switch (length)
-        {
-            case 0:
-            case 1:
-                return false;
-            case 2:
-                return arg[0] == '-';
-            default:
-                return arg[0] == '-' && arg[1] == '-';
         }
-    }
 
-    std::string extract_full_arg(const char* arg)
-    {
-        size_t length = strlen(arg);
+        if (new_spec == NULL)
+            error("unexpected argument %s", current_arg);
 
-        switch (length)
-        {
-            case 0:
-            case 1:
-                return {};
-            case 2:
-                if (arg[0] != '-')
-                    return {};
-                return arg_spec_of(arg[1]).as_full;
-            default:
-                if (arg[0] != '-' || arg[1] != '-')
-                    return {};
-                return arg_spec_of(arg + 2).as_full;
+        active_spec = new_spec;
+        if (active_spec->type == ARG_HELP) {
+            print_help(parser);
+            exit(1);
         }
-    }
 
-private:
-    help_callback m_help_callback;
-    std::vector<ArgSpec> m_args;
-    size_t m_num_positional_args;
-    std::unordered_map<std::string, std::vector<std::string>> m_parsed_args;
-};
+        active_spec->parsed = true;
+    }
+}
+
+static inline bool is_set(const arg_spec_t *spec)
+{
+    return spec->parsed;
+}
+
+static inline const char *get(const arg_spec_t *spec)
+{
+    if (spec->values.count == 0)
+        error("no argument provided for %s", spec->as_full);
+
+    return spec->values.blobs[0].data;
+}
+
+static inline uint64_t get_uint(const arg_spec_t *spec)
+{
+    return strtoull(get(spec), NULL, 10);
+}
+
+static inline uint64_t get_uint_or(
+    const arg_spec_t *spec, uint64_t default_value
+)
+{
+    if (is_set(spec))
+        return get_uint(spec);
+
+    return default_value;
+}
